Passage des données en utilisant une structure

+++ src/mediaModule/text.cpp
@@ -31,7 +31,7 @@
 #include "mediaModule/text.hpp"
 
 
-Text::Text(const std::string &_name, const std::string &_text, int _altitude, int _azimuth , s_font* _myFont, const TEXT_ALIGN &_textAlign,  const Vec3f &color)
+Text::Text(const std::string &_name, const std::string &_text, float _altitude, float _azimuth , s_font* _myFont, const TEXT_ALIGN &_textAlign,  const Vec3f &color)
 {
 	name= _name;
 	text= _text;
